Redesign, completion of mainpage links, refactoring of models, views accross the applicaiton
