+++ requirements.txt
@@ -1,0 +1,17 @@
+asgiref==3.5.2
+astroid==2.12.12
+crispy-bootstrap5==0.7
+dill==0.3.6
+Django==4.1.2
+django-crispy-forms==1.14.0
+isort==5.10.1
+lazy-object-proxy==1.8.0
+mccabe==0.7.0
+Pillow==9.3.0
+platformdirs==2.5.2
+pybind11==2.10.0
+pylint==2.15.5
+sqlparse==0.4.3
+tomli==2.0.1
+tomlkit==0.11.6
+wrapt==1.14.1
