+++ src/bookview/templates/bookview/book_add.upd.html
@@ -11,7 +11,7 @@
                 <form action="" method="post" enctype="multipart/form-data">
                     {% csrf_token %}
                     {{ form|crispy }}
-                    <input type="submit" class="btn btn-success" value="{{ operation }}">
+                    <input type="submit" class="btn btn-outline-success mt-auto" value="{{ operation }}">
 
 
                 </form>
