+++ src/bookview/templates/bookview/booklist.html
@@ -31,7 +31,7 @@ ADMIN: All Books
     {% for obj in object_list %}
 
     <tr>
-        <td><a href="{% url 'bookview:upd_book' pk=obj.pk %}"><button type="button" class="btn btn-warning"><b><i class="bi bi-pencil"></i>{{ obj.pk }}</button></b></button></a></td>
+        <td><a href="{% url 'bookview:upd_book' pk=obj.pk %}"><button type="button" class="btn btn-outline-warning mt-auto"><b><i class="bi bi-pencil"></i>{{ obj.pk }}</button></b></button></a></td>
         <td><img src="{{ obj.book_image.url }}" style="width: 6rem;"></td>
         <td><a href="{% url 'bookview:details_book' pk=obj.pk %}">{{ obj.book_name }}</a></td>
         <td>{{ obj.book_aut }}</td>
@@ -44,7 +44,7 @@ ADMIN: All Books
         <td>{{ obj.created_by }}</td>
 
         <td>
-          <a href="{% url 'bookview:del_book' pk=obj.pk %}"><button type="button" class="btn btn-danger"><i class="bi bi-trash"></i></button></a>
+          <a href="{% url 'bookview:del_book' pk=obj.pk %}"><button type="button" class="btn btn-outline-danger mt-auto"><i class="bi bi-trash"></i></button></a>
         </td>
     </tr>
