+++ src/cartview/templates/cartview/cartitem_view.html
@@ -1,0 +1,11 @@
+{% extends "base.html" %}
+{% load crispy_forms_tags %}
+{% block title %}
+Cart
+{% endblock title %}
+
+{% block body %}
+
+
+</body>
+{% endblock body %}
