+++ src/reflib/templates/reflib/add.upd_aut.html
@@ -11,7 +11,7 @@
     <form action="" method="post">
         {% csrf_token %}
         {{ form|crispy }}
-        <input type="submit" class="btn btn-success" value="{{ operation }}">
+        <input type="submit" class="btn btn-outline-success mt-auto" value="{{ operation }}">
 
 
     </form>
