+++ src/reflib/templates/reflib/pub_add.upd.html
@@ -10,7 +10,7 @@
     <form action="" method="post">
         {% csrf_token %}
         {{ form|crispy }}
-        <input type="submit" class="btn btn-success" value="{{ operation }}">
+        <input type="submit" class="btn btn-outline-success mt-auto" value="{{ operation }}">
 
     </form>
 </body>
