+++ src/reflib/templates/reflib/pub_list.html
@@ -24,11 +24,11 @@ All Publishers
     {% for obj in object_list %}
 
     <tr>
-        <td><a href="{% url 'reflib:upd_publisher' pk=obj.pk %}"><button type="button" class="btn btn-warning"><b><i class="bi bi-pencil"></i>{{ obj.pk }}</b></button></a></td>
+        <td><a href="{% url 'reflib:upd_publisher' pk=obj.pk %}"><button type="button" class="btn btn-outline-warning mt-auto"><b><i class="bi bi-pencil"></i>{{ obj.pk }}</b></button></a></td>
         <td>{{ obj.pub_name }}</td>
         <td>{{ obj.pub_country }}</td>
         <td>{{ obj.pub_note }}</td>
-        <td><a href="{% url 'reflib:del_publisher' pk=obj.pk %}"><button type="button" class="btn btn-danger"><b>Delete<i class="bi bi-x-lg"></i></b></button></a></td>
+        <td><a href="{% url 'reflib:del_publisher' pk=obj.pk %}"><button type="button" class="btn btn-outline-danger mt-auto"><i class="bi bi-trash"></i></button></a></td>
     </tr>
     
     {% endfor %}
