+++ src/templates/base.html
@@ -58,9 +58,8 @@
                         </ul>
                         <form class="d-flex">
                             <button class="btn btn-outline-dark" type="submit">
-                                <i class="bi-cart-fill me-1"></i>
-                                Cart
-                                <span class="badge bg-dark text-white ms-1 rounded-pill">0</span>
+                                <i class="bi bi-cart-check"></i>
+                                User Cart
                             </button>
                         </form>
                     </div>
